Slight change in wording and links
It said: "Join us and let's build Stormkit together" on the left with a link with "Get started for free" to the right. These are two different call to actions with two different purposes. Now it is
Join us and let's build Stormkit together with a link to actually coin the community on Discord.
The text below is to get them to get started for free. Happy to discuss.